compatible with nerfview=0.1.0
